+++ src/com/piedpiper/gui/PiedController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.piedpiper;
+package com.piedpiper.gui;
 
 import java.net.URL;
 import java.util.ResourceBundle;
