+++ src/com/piedpiper/gui/PiedPiper.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.piedpiper;
+package com.piedpiper.gui;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
